#11: pick tethering was tested, and none of it could fail a PR
The backlog says "pick tethering untested". It is tested — 15 tests in
tests/api/test_pick_rookie_anchor.py, and they are good ones. The defect
is that none of them run in CI's hard gate.

Measured before the change:

    pytest tests/api/test_pick_rookie_anchor.py -m "not livedata"
    -> 15 deselected, 0 run

tests/conftest.py marks every test in _LIVEDATA_MODULES as ``livedata``,
moving it to a non-blocking advisory tier. That policy is right and
stays: a yahooBoone row-count dip should not block every PR, and the
module-granularity choice is documented as deliberate.

But the granularity is the problem here. The module held two classes:

  TestAnchorEndToEnd   reads exports/latest/dynasty_data_*.json —
                       genuinely data-coupled, correctly advisory
  TestAnchorPassCore   "Synthetic playersArray exercises — no live data
                       required", its own words

The synthetic half guards pipeline step 11, current-year pick
tethering, which sets the rankDerivedValue of all 72 slot picks on the
live board. It cannot be broken by scrape churn and it could not fail a
PR. Same shape as defect #2 (test_anchor_curve_extrapolation_monotone):
a real test nobody is running.

Split into test_pick_rookie_anchor_core.py — 9 tests now block, the 6
end-to-end ones stay advisory. Total unchanged at 15; the suite goes
4199 passed / 325 deselected to 4212 / 316, which is the 9 plus this
change's own 4 guards.

## Two more things the list was hiding

test_footballguys_source.py was listed and does not exist — a dead
exemption that reads as load-bearing. Removed.

tests/test_livedata_policy.py now checks the two things about the policy
that can be checked mechanically: every entry names a real file, and the
rescued module stays rescued. Both observed failing against the pre-fix
conftest.

## Flagged, not fixed

test_dlf_source.py carries TestDlfCsvEnrichment, which builds a
temporary CSV explicitly "without touching the real CSVs/site_raw tree"
— the same pure-logic half sitting in a live-data module. Splitting it
is the same move and wants its own change rather than being smuggled
into one about pick tethering.

Gates: 4212 passed, ruff format --check and ruff check both clean.
